+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-bMZt0UgW.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-CSqmkJo5.js";import{_ as a}from"./mermaid.core-BUzASpOK.js";import"./chunk-5VM5RSS4-CQxkuyYn.js";import"./chunk-XXDRQBXY-Cxo1gRVY.js";import"./chunk-POPQ4Y6H-6v9KJk35.js";import"./chunk-F27PBJKO-DBBPdMUY.js";import"./channel-C4Oos9C2.js";import"./index-BTVclddX.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};